DOMA-742 review fix move test data rename file
