+++ apps/condo/domains/organization/schema/SbbolAuth.test.js
@@ -1,4 +1,4 @@
-const { SbbolOrganization } = require('@condo/domains/organization/sbbol/synch')
+const { SbbolOrganization } = require('@condo/domains/organization/sbbol/sync')
 const { EXAMPLE_USER_INFO, SBBOL_IMPORT_NAME } = require('@condo/domains/organization/sbbol/common')
 const faker = require('faker')
 const { v4: uuid } = require('uuid')
@@ -8,6 +8,33 @@ const { updateTestUser, User: UserAPI } = require('@condo/domains/user/utils/tes
 const { makeClientWithRegisteredOrganization } = require('@condo/domains/organization/utils/testSchema/Organization')
 const { updateTestOrganization, Organization: OrganizationAPI, OrganizationEmployee: OrganizationEmployeeAPI  } = require('@condo/domains/organization/utils/testSchema')
 const { makeLoggedInAdminClient } = require('@core/keystone/test.utils')
+
+const EXAMPLE_USER_INFO = {
+    sub: 'f164e8c81fa7b5cd43d5d03164cf74764b2402d6314c67daa0964d8e691fd543',
+    iss: 'https://edupirfintech.sberbank.ru:9443',
+    inn: '7784523718',
+    orgJuridicalAddress: '121222, Россия, г.Москва, г. Москва, ул.Космонавтов, д.1, корп./стр.1, кв.1',
+    orgFullName: 'Общество с ограниченной ответственностью ТО-Партнер-626-01"',
+    OrgName: 'ООО "ТО-Партнер-626-01"',
+    tbIdentCode: '38',
+    userGuid: 'c429a86d-38fc-3ac3-e054-00144ffb59b5',
+    individualExecutiveAgency: 0,
+    terBank: 'Московский Банк Сбербанка РФ',
+    userSignatureType: 'Единственная подпись',
+    aud: '111286',
+    userCryptoType: 'SMS',
+    userGroups: 'Руководитель',
+    summOfferSmartCredit: 0,
+    orgLawFormShort: 'ООО',
+    HashOrgId: '8c80ef870028888bc444c27ba90873619cc7a6c99febd89f0e9fee155219b752',
+    offerSmartCredit: false,
+    orgOgrn: '1137746216261',
+    isIdentified: false,
+    inquiryOrder: true,
+    phone_number: '79057362611',
+    orgLawForm: 'Общество с ограниченной ответственностью',
+    email: 'Client_Test62611@test.ru',
+}
 
 
 const userInfo = () => {
